refactor: deprecate parse function in favor of Expressions::from_str
