+++ extendr-api/src/lib.rs
@@ -328,6 +328,7 @@ pub use robj::Robj;
 pub use std::convert::{TryFrom, TryInto};
 pub use std::ops::Deref;
 pub use std::ops::DerefMut;
+pub use std::str::FromStr;
 
 #[cfg(feature = "serde")]
 pub mod serializer;
