+++ extendr-api/src/prelude.rs
@@ -4,10 +4,11 @@
 //! using deprecated features.
 
 pub use super::error::Error;
+pub use super::functions::parse;
 pub use super::functions::{
     base_namespace, blank_scalar_string, blank_string, eval_string, eval_string_with_params,
     find_namespace, find_namespaced_function, global_function, na_string, new_env, nil_value,
-    parse, srcref,
+    srcref,
 };
 pub use super::{
     print_r_error, print_r_output, CanBeNA, Rtype, FALSE, NA_INTEGER, NA_LOGICAL, NA_REAL,
@@ -58,6 +59,7 @@ pub use extendr_macros::{
     call, extendr, extendr_module, pairlist, IntoDataFrameRow, IntoList, Rraw, TryFromList, R,
 };
 pub use std::convert::{TryFrom, TryInto};
+pub use std::str::FromStr;
 
 #[cfg(feature = "ndarray")]
 pub use ::ndarray;
